+++ answer.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"fmt"
+)
+
+type Answer struct {
+	example *Example
+	value   int
+}
+
+func (a *Answer) GetAnalysis() string {
+	if a.example.isCorrectAnswer(a.value) {
+		return "Правильно!"
+	}
+	return fmt.Sprintf("Неправильно. Правильный ответ %v", a.example.Answer())
+}